data: rename fiches-sp and fiches-mt
