+++ packages/code-du-travail-data/dataset/fiches_service_public/package.json
@@ -17,11 +17,11 @@
     "xml-js": "^1.6.11"
   },
   "scripts": {
-    "backup": "cp fiches-sp-travail.json fiches-sp-travail.old.json ",
+    "backup": "cp fiches-sp.json fiches-sp.old.json ",
     "diff": "node diff.js",
-    "updateTheme": "node updateTheme.js > out.json && mv out.json fiches-sp-travail.json  && yarn report",
+    "updateTheme": "node updateTheme.js > out.json && mv out.json fiches-sp.json  && yarn report",
     "fetchData": "node fetchData.js",
-    "start": "yarn backup && node index.js > fiches-sp-travail.json",
+    "start": "yarn backup && node index.js > fiches-sp.json",
     "report": "node diff.js && node missingTheme.js",
     "test": "jest"
   }
